+++ src/main/java/micycle/pgs/PGS_Morphology.java
@@ -178,7 +178,7 @@ public final class PGS_Morphology {
 			LineString l = (LineString) g;
 			return toPShape(GaussianLineSmoothing.get(l, Math.max(sigma, 1)));
 		}
-		System.err.println(g.getGeometryType() + " are not supported for this method (yet).");
+		System.err.println(g.getGeometryType() + " are not supported for the smoothGaussian() method (yet).");
 		return shape;
 	}